Ajout message de confirmation pour la suppression d'une fiche de soins

+++ src/main/java/viewPackage/PanneauListingFichesDeSoin.java
@@ -81,23 +81,29 @@ public class PanneauListingFichesDeSoin extends JPanel {
                 }
             }
             else if(event.getSource() == supprimerButton){
-                listeSelectionnee = resultatRecherche.getSelectionModel();
-                int indiceLigneSelectionnee = listeSelectionnee.getMinSelectionIndex();
-                try{
-                    String critere = (String)comboBoxListingFiches.getSelectedItem();
-                    soinsTries = controller.getSoinsTries(critere);
-                    SoinAvance soinASup = soinsTries.get(indiceLigneSelectionnee);
-                    controller.supprimerSoin(soinASup);
-                    JOptionPane.showMessageDialog(null, "Le soin a été correctemen supprimé de la base de données !","Confirmation", JOptionPane.INFORMATION_MESSAGE);
-                } catch (VeterinaireException e) {
-                    JOptionPane.showMessageDialog(null, e.getMessage());
-                } catch (SoinException e) {
-                    JOptionPane.showMessageDialog(null, e.getMessage());
-                } catch (SingletonConnectionException e) {
-                    JOptionPane.showMessageDialog(null, e.getMessage());
-                }
-                catch (Exception exception){
-                    System.out.println("Exception : " + exception.getMessage());
+                JOptionPane confirmation = new JOptionPane();
+                int i = JOptionPane.showConfirmDialog(null, "La suppression est irréversible. Êtes-vous sûr de vouloir continuer?",
+                        "Veuillez confirmer votre choix",
+                        JOptionPane.YES_NO_OPTION);
+                if(i == 0) {
+                    listeSelectionnee = resultatRecherche.getSelectionModel();
+                    int indiceLigneSelectionnee = listeSelectionnee.getMinSelectionIndex();
+                    try {
+                        String critere = (String) comboBoxListingFiches.getSelectedItem();
+                        soinsTries = controller.getSoinsTries(critere);
+                        SoinAvance soinASup = soinsTries.get(indiceLigneSelectionnee);
+                        controller.supprimerSoin(soinASup);
+                        trierButton.doClick();
+                        JOptionPane.showMessageDialog(null, "Le soin a été correctemen supprimé de la base de données !", "Confirmation", JOptionPane.INFORMATION_MESSAGE);
+                    } catch (VeterinaireException e) {
+                        JOptionPane.showMessageDialog(null, e.getMessage());
+                    } catch (SoinException e) {
+                        JOptionPane.showMessageDialog(null, e.getMessage());
+                    } catch (SingletonConnectionException e) {
+                        JOptionPane.showMessageDialog(null, e.getMessage());
+                    } catch (Exception exception) {
+                        System.out.println("Exception : " + exception.getMessage());
+                    }
                 }
             }
             else if(event.getSource() == modifierButton){
@@ -122,6 +128,4 @@ public class PanneauListingFichesDeSoin extends JPanel {
             }
         }
     }
-
-
 }
